requirement.txt
mysql.connector-python==9.2.0